fix : ãFK 필드 맵핑 수정 및 ãPaymentValidationService 추가 및 WalletService 로직 수정

+++ backend/src/main/java/com/example/pproject/resume/service/ResumeService.java
@@ -25,7 +25,7 @@ public class ResumeService {
     private final JobApplicationRepository jobApplicationRepository;
 
     // 1. 내 이력서 목록 조회
-    public List<ResumeResponse> getResumes(Integer userId) {
+    public List<ResumeResponse> getResumes(Long userId) {
         return resumeRepository.findAllByUserId(userId).stream()
                 .map(ResumeResponse::from)
                 .collect(Collectors.toList());
@@ -39,7 +39,7 @@ public class ResumeService {
 
     // 3. 이력서 생성
     @Transactional
-    public Long createResume(ResumeRequest request, Integer userId) {
+    public Long createResume(ResumeRequest request, Long userId) {
         UserEntity user = getUser(userId);
         Resume resume = request.toEntity(user);
 
@@ -58,7 +58,7 @@ public class ResumeService {
 
     // 4. 이력서 수정
     @Transactional
-    public Long updateResume(Long resumeId, ResumeRequest request, Integer userId) {
+    public Long updateResume(Long resumeId, ResumeRequest request, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
 
@@ -81,7 +81,7 @@ public class ResumeService {
 
     // 5. 이력서 삭제
     @Transactional
-    public void deleteResume(Long resumeId, Integer userId) {
+    public void deleteResume(Long resumeId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
 
@@ -96,7 +96,7 @@ public class ResumeService {
 
     // 6. 이력서 복제
     @Transactional
-    public Long copyResume(Long resumeId, Integer userId) {
+    public Long copyResume(Long resumeId, Long userId) {
         Resume original = getResumeEntity(resumeId);
         validateOwner(original, userId);
 
@@ -130,61 +130,61 @@ public class ResumeService {
 
     // 하위 항목 관리
     @Transactional
-    public void addCareer(Long resumeId, ResumeRequest.CareerDto dto, Integer userId) {
+    public void addCareer(Long resumeId, ResumeRequest.CareerDto dto, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getCareers().add(dto.toEntity(resume));
     }
     @Transactional
-    public void deleteCareer(Long resumeId, Long careerId, Integer userId) {
+    public void deleteCareer(Long resumeId, Long careerId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getCareers().removeIf(c -> c.getId().equals(careerId));
     }
     @Transactional
-    public void addProject(Long resumeId, ResumeRequest.ProjectDto dto, Integer userId) {
+    public void addProject(Long resumeId, ResumeRequest.ProjectDto dto, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getProjects().add(dto.toEntity(resume));
     }
     @Transactional
-    public void deleteProject(Long resumeId, Long projectId, Integer userId) {
+    public void deleteProject(Long resumeId, Long projectId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getProjects().removeIf(p -> p.getId().equals(projectId));
     }
     @Transactional
-    public void addCertificate(Long resumeId, ResumeRequest.CertificateDto dto, Integer userId) {
+    public void addCertificate(Long resumeId, ResumeRequest.CertificateDto dto, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getCertificates().add(dto.toEntity(resume));
     }
     @Transactional
-    public void deleteCertificate(Long resumeId, Long certId, Integer userId) {
+    public void deleteCertificate(Long resumeId, Long certId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getCertificates().removeIf(c -> c.getId().equals(certId));
     }
     @Transactional
-    public void addLink(Long resumeId, ResumeRequest.LinkDto dto, Integer userId) {
+    public void addLink(Long resumeId, ResumeRequest.LinkDto dto, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getLinks().add(dto.toEntity(resume));
     }
     @Transactional
-    public void deleteLink(Long resumeId, Long linkId, Integer userId) {
+    public void deleteLink(Long resumeId, Long linkId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getLinks().removeIf(l -> l.getId().equals(linkId));
     }
     @Transactional
-    public void addAttachment(Long resumeId, ResumeRequest.AttachmentDto dto, Integer userId) {
+    public void addAttachment(Long resumeId, ResumeRequest.AttachmentDto dto, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getAttachments().add(dto.toEntity(resume));
     }
     @Transactional
-    public void deleteAttachment(Long resumeId, Long attachmentId, Integer userId) {
+    public void deleteAttachment(Long resumeId, Long attachmentId, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
         resume.getAttachments().removeIf(a -> a.getId().equals(attachmentId));
@@ -192,7 +192,7 @@ public class ResumeService {
 
     // 프로필 사진 업데이트
     @Transactional
-    public void updateProfileImage(Long resumeId, String photoUrl, Integer userId) {
+    public void updateProfileImage(Long resumeId, String photoUrl, Long userId) {
         Resume resume = getResumeEntity(resumeId);
         validateOwner(resume, userId);
 
@@ -210,7 +210,7 @@ public class ResumeService {
         }
     }
 
-    private UserEntity getUser(Integer userId) {
+    private UserEntity getUser(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found: " + userId));
     }
@@ -218,7 +218,7 @@ public class ResumeService {
         return resumeRepository.findById(resumeId)
                 .orElseThrow(() -> new IllegalArgumentException("Resume not found: " + resumeId));
     }
-    private void validateOwner(Resume resume, Integer userId) {
+    private void validateOwner(Resume resume, Long userId) {
         if (!resume.getUser().getId().equals(userId)) {
             throw new IllegalArgumentException("Unauthorized access to resume");
         }
